Add support for DID attributes
Implemented as a stream of events, as per EIP-1056.
By applying all changes to the attributes of a user, the currently valid
set of attributes can be obtained.

This greatly reduces the complexity of the smart contract.
As the attributes are only read from off-chain applications, this is
feasible.

See: https://github.com/ethereum/EIPs/issues/1056
See: https://eips.ethereum.org/EIPS/eip-1056